+++ src/main/kotlin/com/zepben/evolve/streaming/get/CimConsumerClient.kt
@@ -90,16 +90,8 @@ abstract class CimConsumerClient<T : BaseService, U : BaseProtoToCim>(executor: 
      * - When [GrpcResult.wasFailure], the error that occurred retrieving or processing the object, accessible via [GrpcResult.thrown].
      * Note the [CimConsumerClient] warning in this case.
      */
-    fun getIdentifiedObjects(mRIDs: Sequence<String>): GrpcResult<MultiObjectResult> = tryRpc {
-        val results = mutableMapOf<String, IdentifiedObject>()
-        val failed = mRIDs.toMutableSet()
-        processIdentifiedObjects(mRIDs).forEach { result ->
-            result.identifiedObject?.let {
-                results[it.mRID] = it
-                failed.remove(it.mRID)
-            }
-        }
-        MultiObjectResult(results, failed)
+    fun getIdentifiedObjects(mRIDs: Sequence<String>): GrpcResult<MultiObjectResult> = handleMultiObjectRPC(mRIDs) {
+        processIdentifiedObjects(mRIDs)
     }
 
     protected abstract fun processIdentifiedObjects(mRIDs: Sequence<String>): Sequence<ExtractResult>
@@ -120,12 +112,19 @@ abstract class CimConsumerClient<T : BaseService, U : BaseProtoToCim>(executor: 
     protected inline fun <reified CIM : IdentifiedObject> getOrAdd(mRID: String, addFromPb: U.() -> CIM?): CIM? =
         service.get(CIM::class, mRID) ?: protoToCim.addFromPb()
 
-    protected fun handleMultiObjectRPC(processor: () -> Sequence<ExtractResult>): GrpcResult<MultiObjectResult> =
+    /**
+     * @param processor The function which returns results to be handled.
+     * @param mRIDs The mRIDs expected to be retrieved from a call to [processor]
+     */
+    protected fun handleMultiObjectRPC(mRIDs: Sequence<String>? = null, processor: () -> Sequence<ExtractResult>): GrpcResult<MultiObjectResult> =
         tryRpc {
             val results = mutableMapOf<String, IdentifiedObject>()
-            val failed = mutableSetOf<String>()
+            val failed = mRIDs?.toMutableSet() ?: mutableSetOf()
             processor().forEach { result ->
-                result.identifiedObject?.let { results[it.mRID] = it } ?: failed.add(result.mRID)
+                result.identifiedObject?.let {
+                    results[it.mRID] = it
+                    failed.remove(it.mRID)
+                } ?: failed.add(result.mRID)
             }
             MultiObjectResult(results, failed)
         }