+++ src/test/kotlin/com/zepben/evolve/streaming/get/testservices/TestDiagramConsumerService.kt
@@ -8,16 +8,18 @@
 
 package com.zepben.evolve.streaming.get.testservices
 
-import com.zepben.protobuf.dc.DiagramConsumerGrpc
-import com.zepben.protobuf.dc.GetIdentifiedObjectsRequest
-import com.zepben.protobuf.dc.GetIdentifiedObjectsResponse
+import com.zepben.protobuf.dc.*
 import io.grpc.stub.StreamObserver
 
 class TestDiagramConsumerService : DiagramConsumerGrpc.DiagramConsumerImplBase() {
 
     lateinit var onGetIdentifiedObjects: (request: GetIdentifiedObjectsRequest, response: StreamObserver<GetIdentifiedObjectsResponse>) -> Unit
+    lateinit var onGetDiagramObjects: (request: GetDiagramObjectsRequest, response: StreamObserver<GetDiagramObjectsResponse>) -> Unit
 
     override fun getIdentifiedObjects(response: StreamObserver<GetIdentifiedObjectsResponse>): StreamObserver<GetIdentifiedObjectsRequest> =
         TestStreamObserver(response, onGetIdentifiedObjects)
+
+    override fun getDiagramObjects(responseObserver: StreamObserver<GetDiagramObjectsResponse>?): StreamObserver<GetDiagramObjectsRequest> =
+        TestStreamObserver(responseObserver!!, onGetDiagramObjects)
 
 }